move `spHandlerOpen` to hub from serialport.go
