+++ hub.go
@@ -16,6 +16,7 @@
 package main
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"html"
@@ -30,6 +31,7 @@ import (
 	"github.com/arduino/arduino-create-agent/tools"
 	"github.com/arduino/arduino-create-agent/upload"
 	log "github.com/sirupsen/logrus"
+	"go.bug.st/serial"
 )
 
 type hub struct {
@@ -278,6 +280,90 @@ func (h *hub) checkCmd(m []byte) {
 	} else {
 		go h.spErr("Could not understand command.")
 	}
+}
+
+func (h *hub) spHandlerOpen(portname string, baud int, buftype string) {
+
+	log.Print("Inside spHandler")
+
+	var out bytes.Buffer
+
+	out.WriteString("Opening serial port ")
+	out.WriteString(portname)
+	out.WriteString(" at ")
+	out.WriteString(strconv.Itoa(baud))
+	out.WriteString(" baud")
+	log.Print(out.String())
+
+	conf := &SerialConfig{Name: portname, Baud: baud, RtsOn: true}
+
+	mode := &serial.Mode{
+		BaudRate: baud,
+	}
+
+	sp, err := serial.Open(portname, mode)
+	log.Print("Just tried to open port")
+	if err != nil {
+		//log.Fatal(err)
+		log.Print("Error opening port " + err.Error())
+		//h.broadcastSys <- []byte("Error opening port. " + err.Error())
+		h.broadcastSys <- []byte("{\"Cmd\":\"OpenFail\",\"Desc\":\"Error opening port. " + err.Error() + "\",\"Port\":\"" + conf.Name + "\",\"Baud\":" + strconv.Itoa(conf.Baud) + "}")
+
+		return
+	}
+	log.Print("Opened port successfully")
+	//p := &serport{send: make(chan []byte, 256), portConf: conf, portIo: sp}
+	// we can go up to 256,000 lines of gcode in the buffer
+	p := &serport{
+		sendBuffered: make(chan string, 256000),
+		sendNoBuf:    make(chan []byte),
+		sendRaw:      make(chan string),
+		portConf:     conf,
+		portIo:       sp,
+		portName:     portname,
+		BufferType:   buftype,
+	}
+
+	p.OnMessage = func(msg []byte) {
+		h.broadcastSys <- msg
+	}
+	p.OnClose = func(port *serport) {
+		h.serialPortList.MarkPortAsClosed(p.portName)
+		h.serialPortList.List()
+	}
+
+	var bw Bufferflow
+
+	switch buftype {
+	case "timed":
+		bw = NewBufferflowTimed(portname, h.broadcastSys)
+	case "timedraw":
+		bw = NewBufferflowTimedRaw(portname, h.broadcastSys)
+	case "default":
+		bw = NewBufferflowDefault(portname, h.broadcastSys)
+	default:
+		log.Panicf("unknown buffer type: %s", buftype)
+	}
+
+	bw.Init()
+	p.bufferwatcher = bw
+
+	h.serialHub.Register(p)
+	defer h.serialHub.Unregister(p)
+
+	h.serialPortList.MarkPortAsOpened(portname)
+	h.serialPortList.List()
+
+	// this is internally buffered thread to not send to serial port if blocked
+	go p.writerBuffered()
+	// this is thread to send to serial port regardless of block
+	go p.writerNoBuf()
+	// this is thread to send to serial port but with base64 decoding
+	go p.writerRaw()
+
+	p.reader(buftype)
+
+	h.serialPortList.List()
 }
 
 type logWriter struct {
